+++ app/src/docs/scripts/search/indexer.test.js
@@ -1,0 +1,8 @@
+import { processModuleReferenceDocs, processMdxDocsPages } from './utils'
+
+describe('Indexer', async () => {
+    it('loads the correct number of modules', async () => {
+        const mods = Object.keys(await processModuleReferenceDocs())
+        expect(mods.length).toEqual(170)
+    })
+})
